Tidy

+++ routes/edit.js
@@ -2,32 +2,35 @@ const express = require('express');
 const router = express.Router();
 const fs = require('fs').promises;
 const path = require('path');
+const sharp = require('sharp');
 let data = require('../data/data.json');
 const multer  = require('multer');
 const imageUpload = multer({ dest: 'uploads/' });
 
-const storage = multer.diskStorage({
-	destination: function (req, file, cb) {
-		if (file.fieldname === 'image') {
-			cb(null, 'public/images/');
-		}
-		else if (file.fieldname === 'thumbnail') {
-			cb(null, 'public/images/thumbnails/');
-		} else if (file.fieldname.includes('video')) {
-			cb(null, 'public/videos/');
-		} else if (file.fieldname === 'galleryImage') {
-			cb(null, 'public/images/gallery');
-		} else if (file.fieldname === 'galleryThumbnail') {
-			cb(null, 'public/images/gallery/thumbnails');
-		} else {
-			cb(null, 'public/video/');
-		}
-	},
-	filename: function (req, file, cb) {
-		const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
-		cb(null, `${file.fieldname}-${uniqueSuffix}${path.extname(file.originalname)}`);
-	}
-});
+// const storage = multer.diskStorage({
+// 	destination: function (req, file, cb) {
+// 		if (file.fieldname === 'image') {
+// 			cb(null, 'public/images/');
+// 		}
+// 		else if (file.fieldname === 'thumbnail') {
+// 			cb(null, 'public/images/thumbnails/');
+// 		} else if (file.fieldname.includes('video')) {
+// 			cb(null, 'public/videos/');
+// 		} else if (file.fieldname === 'galleryImage') {
+// 			cb(null, 'public/images/gallery');
+// 		} else if (file.fieldname === 'galleryThumbnail') {
+// 			cb(null, 'public/images/gallery/thumbnails');
+// 		} else {
+// 			cb(null, 'public/video/');
+// 		}
+// 	},
+// 	filename: function (req, file, cb) {
+// 		const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
+// 		cb(null, `${file.fieldname}-${uniqueSuffix}${path.extname(file.originalname)}`);
+// 	}
+// });
+
+const storage = multer.memoryStorage();
 
 const upload = multer({storage: storage});
 
@@ -62,16 +65,8 @@ router.post('/edit', function(req, res, next) {
 				deleteFile(`./public/videos/${data.headerVideoWebm}`);
 				newData.headerVideoWebm = req.files.videoWebm[0].filename;
 			}
-			// Update general site metadata
-			const writeableState = Object.keys(newData).filter(x => {
-				if (typeof newData[x] !== 'object') {
-					return true;
-				}
-			})
-			writeableState.forEach(x => {
-				if (req.body[x]) {
-					newData[x] = req.body[x];
-				}
+			Object.keys(req.body).forEach(x => {
+				newData[x] = req.body[x];
 			});
 
 			// Update Socials
@@ -101,33 +96,98 @@ router.post('/edit/:path', function(req, res, next) {
 		} else {
 			const portfolioIndex = req.params.path;
 			const newData = { ...data };
-			const writeableState = Object.keys(newData.portfolioItems[portfolioIndex]).filter(x => {
-				if (typeof newData.portfolioItems[portfolioIndex][x] !== 'object') {
-					return true;
-				}
-			})
-			writeableState.forEach(x => {
-				if (req.body[x]) {
-					newData.portfolioItems[portfolioIndex][x] = req.body[x];
-				}
+			const promises = [];
+			Object.keys(req.body).forEach(x => {
+				newData.portfolioItems[portfolioIndex][x] = req.body[x];
 			});
 			if (req.files.thumbnail) {
-				deleteFile(`./public/images/thumbnails/${data.portfolioItems[portfolioIndex].thumbnail}`);
-				newData.portfolioItems[portfolioIndex].thumbnail = req.files.thumbnail[0].filename;
+				const path = 'public/images/thumbnails/';
+				const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
+				const fileName = `${req.files.thumbnail[0].fieldname}-${uniqueSuffix}.webp`;
+				promises.push(new Promise((resolve, reject) => {
+					sharp(req.files.thumbnail[0].buffer)
+						.webp({
+							quality: 100,
+						})
+						.resize({
+							width: 300 * Number(data.portfolioItems[portfolioIndex].gridLength),
+							height: 300,
+							fit: 'cover'
+						})
+						.toFile(`${path}${fileName}`)
+						.then(info => {
+							deleteFile(`./public/images/thumbnails/${data.portfolioItems[portfolioIndex].thumbnail}`);
+							newData.portfolioItems[portfolioIndex].thumbnail = fileName;
+							resolve();
+						});
+				}));
 			}
 			if (req.files.image) {
-				deleteFile(`./public/images/${data.portfolioItems[portfolioIndex].image}`);
-				newData.portfolioItems[portfolioIndex].image = req.files.image[0].filename;
+				const path = 'public/images/';
+				const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
+				const fileName = `${req.files.image[0].fieldname}-${uniqueSuffix}.webp`;
+				promises.push(new Promise((resolve, reject) => {
+					sharp(req.files.image[0].buffer)
+						.webp({
+							quality: 100
+						})
+						.resize({
+							width: 1280,
+							height: 720,
+							fit: 'cover'
+						})
+						.toFile(`${path}${fileName}`)
+						.then(info => {
+							deleteFile(`./public/images/${data.portfolioItems[portfolioIndex].image}`);
+							newData.portfolioItems[portfolioIndex].image = fileName;
+							resolve();
+						});
+				}));
 			}
 			if (req.files.galleryImage) {
+				const path = 'public/images/gallery/';
+				const pathThumbnail = 'public/images/gallery/thumbnails/';
 				req.files.galleryImage.forEach((val, index) => {
-					newData.portfolioItems[portfolioIndex].gallery.push({
-						image: val.filename,
-						thumbnail: req.files.galleryThumbnail[index].filename
-					});
+					const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
+					const fileName = `${val.fieldname}-${uniqueSuffix}.webp`;
+					const galleryPromises = [];
+					promises.push(new Promise((resolve, reject) => {
+						sharp(val.buffer)
+							.webp({
+								quality: 100
+							})
+							.toFile(`${path}${fileName}`)
+							.then(info => {
+								resolve();
+							});
+					}));
+					promises.push(new Promise((resolve, reject) => {
+						sharp(req.files.galleryThumbnail[index].buffer)
+							.webp({
+								quality: 100
+							})
+							.resize({
+								width: 300,
+								height: 300,
+								fit: 'cover'
+							})
+							.toFile(`${pathThumbnail}${fileName}`)
+							.then(info => {
+								newData.portfolioItems[portfolioIndex].gallery.push({
+									id: Math.round(Math.random() * 1E9),
+									image: fileName,
+									thumbnail: fileName
+								});
+								resolve();
+							});
+					}));
 				})
 			}
-			saveData(newData, res, `/edit/${portfolioIndex}`)
+			Promise.all(promises)
+			.then(() => {
+				console.log(newData.portfolioItems[portfolioIndex].gallery)
+				saveData(newData, res, `/edit/${portfolioIndex}`)
+			})
 		}
 	})
 });
@@ -144,13 +204,18 @@ router.get('/edit/:path', function(req, res, next) {
 	);
 });
 
-router.delete('/edit/:path/:galleryIndex', function(req, res, next) {
-	const galleryItem = data.portfolioItems[req.params.path].gallery[req.params.galleryIndex];
-	deleteFile(`./public/images/gallery/${galleryItem.image}`);
-	deleteFile(`./public/images/gallery/thumbnail/${galleryItem.image}`);
-	const newData = { ...data };
-	newData.portfolioItems[req.params.path].gallery.splice(req.params.galleryIndex, 1);
-	saveData(newData);
+router.delete('/edit/:portfolioIndex/:galleryId', function(req, res, next) {
+	console.log(req.params.galleryId)
+	data.portfolioItems[req.params.portfolioIndex].gallery.forEach((val, index) => {
+		if (val.id.toString() === req.params.galleryId.toString()) {
+			const galleryItem = data.portfolioItems[req.params.portfolioIndex].gallery[index];
+			deleteFile(`./public/images/gallery/${galleryItem.image}`);
+			deleteFile(`./public/images/gallery/thumbnails/${galleryItem.image}`);
+			const newData = { ...data };
+			newData.portfolioItems[req.params.portfolioIndex].gallery.splice(index, 1);
+			saveData(newData);
+		}
+	});
 	res.status(204).send(null);
 });
